% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_package_content.R
\name{package_d1}
\alias{package_d1}
\title{When I consider how my light is spent
Ere half my days in this dark world and wide,
And that one talent which is death to hide
Lodg’d with me useless, though my soul more bent
To serve therewith my Maker, and present
My true account, lest he returning chide,
“Doth God exact day-labour, light denied?”
I fondly ask. But Patience, to prevent
That murmur, soon replies: “God doth not need
Either man’s work or his own gifts: who best
Bear his mild yoke, they serve him best. His state
Is kingly; thousands at his bidding speed
And post o’er land and ocean without rest:
They also serve who only stand and wait.”}
\usage{
package_d1(a, b)
}
\arguments{
\item{a}{The best input ever}

\item{b}{The awesome second input}
}
\description{
When I consider how my light is spent
Ere half my days in this dark world and wide,
And that one talent which is death to hide
Lodg’d with me useless, though my soul more bent
To serve therewith my Maker, and present
My true account, lest he returning chide,
“Doth God exact day-labour, light denied?”
I fondly ask. But Patience, to prevent
That murmur, soon replies: “God doth not need
Either man’s work or his own gifts: who best
Bear his mild yoke, they serve him best. His state
Is kingly; thousands at his bidding speed
And post o’er land and ocean without rest:
They also serve who only stand and wait.”
}
