% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_package_content.R
\name{what_version_are_you}
\alias{what_version_are_you}
\title{Just some example function.
But this function will also be present in another package.}
\usage{
what_version_are_you()
}
\description{
Just some example function.
But this function will also be present in another package.
}
